refactor: extract backup and restore commands
