+++ src/checkers/mod.rs
@@ -5,7 +5,11 @@
 
 pub mod history;
 
+use std::collections::HashMap;
+use crate::checkers::history::last_paths::{LastPaths, Error as LastPathsError};
+use crate::checkers::history::operation::{HoardOperation, Error as HoardOperationError};
 use crate::hoard::{Direction, Hoard};
+use thiserror::Error;
 
 /// Trait for validating [`Hoard`]s.
 ///
@@ -33,4 +37,67 @@ pub trait Checker: Sized {
     ///
     /// Generally, any I/O errors that occur while persisting data.
     fn commit_to_disk(self) -> Result<(), Self::Error>;
+}
+
+/// Errors that may occur while using [`Checkers`].
+#[derive(Debug, Error)]
+pub enum Error {
+    /// An error occurred while comparing paths for this run to the previous one.
+    #[error("error while comparing previous run to current run: {0}")]
+    LastPaths(#[from] LastPathsError),
+    /// An error occurred while checking against remote operations.
+    #[error("error while checking against recent remote operations: {0}")]
+    Operation(#[from] HoardOperationError),
+}
+
+pub(crate) struct Checkers {
+    last_paths: HashMap<String, LastPaths>,
+    operations: HashMap<String, HoardOperation>,
+}
+
+impl Checkers {
+    #[allow(single_use_lifetimes)]
+    pub(crate) fn new<'a, S: AsRef<str>>(hoards: impl IntoIterator<Item=(S, &'a Hoard)>, direction: Direction) -> Result<Self, Error> {
+        let mut last_paths = HashMap::new();
+        let mut operations = HashMap::new();
+
+        for (name, hoard) in hoards {
+            let name = name.as_ref();
+            let lp = LastPaths::new(name, hoard, direction)?;
+            let op = HoardOperation::new(name, hoard, direction)?;
+            last_paths.insert(name.to_string(), lp);
+            operations.insert(name.to_string(), op);
+        }
+
+        Ok(Self {
+            last_paths,
+            operations,
+        })
+    }
+
+    pub(crate) fn check(&mut self) -> Result<(), Error> {
+        let _span = tracing::info_span!("running_checks").entered();
+        for last_path in &mut self.last_paths.values_mut() {
+            last_path.check()?;
+        }
+        for operation in self.operations.values_mut() {
+            operation.check()?;
+        }
+        Ok(())
+    }
+
+    pub(crate) fn commit_to_disk(self) -> Result<(), Error> {
+        let Self {
+            last_paths,
+            operations,
+            ..
+        } = self;
+        for (_, last_path) in last_paths {
+            last_path.commit_to_disk()?;
+        }
+        for (_, operation) in operations {
+            operation.commit_to_disk()?;
+        }
+        Ok(())
+    }
 }
