+++ src/command/mod.rs
@@ -1,5 +1,6 @@
 //! See [`Command`].
 
+mod backup_restore;
 mod cleanup;
 mod diff;
 mod edit;
@@ -9,6 +10,7 @@ mod status;
 use structopt::StructOpt;
 use thiserror::Error;
 
+pub(crate) use backup_restore::{run_backup, run_restore};
 pub(crate) use cleanup::run_cleanup;
 pub(crate) use edit::run_edit;
 pub(crate) use diff::run_diff;
@@ -21,6 +23,18 @@ pub enum Error {
     /// Error occurred while printing the help message.
     #[error("error while printing help message: {0}")]
     PrintHelp(#[from] structopt::clap::Error),
+    /// Error occurred while backing up a hoard.
+    #[error("failed to back up {name}: {error}")]
+    Backup {
+        /// The name of the hoard that failed to back up.
+        name: String,
+        /// The error that occurred.
+        #[source]
+        error: crate::hoard::Error,
+    },
+    /// Error occurred while running [`Checkers`](crate::checkers::Checkers).
+    #[error("error while running or saving consistency checks: {0}")]
+    Checkers(#[from] crate::checkers::Error),
     /// An error occurred while running the cleanup command.
     #[error("error after cleaning up {success_count} log files: {error}")]
     Cleanup {
@@ -36,6 +50,15 @@ pub enum Error {
     /// Error occurred while running the edit command.
     #[error("error while running hoard edit: {0}")]
     Edit(#[from] edit::Error),
+    /// Error occurred while restoring a hoard.
+    #[error("failed to back up {name}: {error}")]
+    Restore {
+        /// The name of the hoard that failed to restore.
+        name: String,
+        /// The error that occurred.
+        #[source]
+        error: crate::hoard::Error,
+    },
     /// Error occurred while running the status command.
     #[error("error while running hoard status: {0}")]
     Status(#[source] crate::hoard::iter::Error),
